+++ CHANGES.md
@@ -5,6 +5,8 @@ PDF-Builder Changes
 
 * Adds the capability to set error correction level for QR-codes.
 
+* Adding QA tests for QR code rendere (QRGen).
+
 ## v0.3.1 (Release on 08.02.2016)
 
 ### Change Request: